refactor repository structure
