--- src/pages/FriendsPage/FriendsPage.tsx
@@ -1,7 +1,0 @@
-import FriendsPageModule from '../../modules/FriendsPage/components/FriendsPageModule/FriendsPageModule'
-
-export default function FriendsPage() {
-  return (
-    <FriendsPageModule />
-  )
-}
