--- src/pages/MainPage/MainPage.tsx
@@ -1,11 +1,0 @@
-import MainPageModule from "../../modules/MainPage/components/MainPageModule";
-
-const MainPage = () => {
-  return (
-    <>
-      <MainPageModule />
-    </>
-  );
-};
-
-export default MainPage;
