--- src/pages/MiningPage/MiningPage.tsx
@@ -1,7 +1,0 @@
-import MiningPageModule from "../../modules/MiningPage/components/MiningPageModule/MiningPageModule";
-
-export default function MiningPage() {
-  return (
-    <MiningPageModule />
-  )
-}
