--- src/pages/TasksPage/TasksPage.tsx
@@ -1,7 +1,0 @@
-import TasksPageModule from "../../modules/TasksPage/components/TasksPageModule/TasksPageModule";
-
-export default function TasksPage() {
-  return (
-    <TasksPageModule />
-  )
-}
