--- src/pages/WelcomePage/WelcomePage.tsx
@@ -1,11 +1,0 @@
-import WelcomePageModule from "../../modules/WelcomePage/components/WelcomePageModule";
-
-const WelcomePage = () => {
-  return (
-    <>
-      <WelcomePageModule />
-    </>
-  );
-};
-
-export default WelcomePage;
