+++ src/pages/friends/friends.page.tsx
@@ -1,0 +1,5 @@
+import FriendsPageModule from "./modules/friendsPage.module";
+
+export default function FriendsPage() {
+    return <FriendsPageModule />;
+}
