+++ src/pages/friends/modules/components/FriendItem/FriendItem.styled.tsx
