+++ src/pages/friends/modules/components/FriendsPageModule/FriendsPageModule.styled.tsx
