+++ src/pages/friends/modules/components/OfferItem/OfferItem.styled.tsx
