+++ src/pages/main/main.page.tsx
@@ -1,0 +1,11 @@
+import MainPageModule from "./modules/mainPage.module";
+
+const MainPage = () => {
+    return (
+        <>
+            <MainPageModule />
+        </>
+    );
+};
+
+export default MainPage;
