+++ src/pages/main/modules/components/MainPageModule.styled.ts
