+++ src/pages/main/modules/components/Tabs/TabsModule.styled.ts
