+++ src/pages/mining/mining.page.tsx
@@ -1,0 +1,7 @@
+import MiningPageModule from "./modules/miningPage.module";
+
+export default function MiningPage() {
+  return (
+    <MiningPageModule />
+  )
+}
