+++ src/pages/mining/modules/components/MiningPageModule/MiningPageModule.styled.tsx
