+++ src/pages/tasks/modules/taskPage.module.tsx
@@ -1,0 +1,20 @@
+import * as Styled from "./components/TasksPageModule/TasksPageModule.styled";
+import { tasksList } from "./data/data";
+import TaskItem from "./taskItem.module";
+
+export default function TasksPageModule() {
+    return (
+        <Styled.Container>
+            <Styled.TasksList>
+                {tasksList.map(({ title, count, status }, index) => (
+                    <TaskItem
+                        key={index}
+                        title={title}
+                        count={count}
+                        status={status}
+                    />
+                ))}
+            </Styled.TasksList>
+        </Styled.Container>
+    );
+}
