+++ src/pages/tasks/tasks.page.tsx
@@ -1,0 +1,7 @@
+import TasksPageModule from "./modules/taskPage.module";
+
+export default function TasksPage() {
+  return (
+    <TasksPageModule />
+  )
+}
