+++ src/pages/welcome/welcome.page.tsx
@@ -1,0 +1,11 @@
+import WelcomePageModule from "./modules/welcomePage.module";
+
+const WelcomePage = () => {
+  return (
+    <>
+      <WelcomePageModule />
+    </>
+  );
+};
+
+export default WelcomePage;
